+++ vite.config.js
@@ -5,6 +5,9 @@ import path from 'path'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    port: Number(process.env.PORT) || 5183,
+  },
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src')